+++ CHANGELOG.rst
@@ -16,6 +16,14 @@ are used for versioning (schema follows below):
 - All backwards incompatible changes are mentioned in this document.
 
 
+0.6.9
+-----
+2016-09-08
+
+- Fixed changing order of the `FormElement`, `FormElementEntry`, `FormHandler`
+  and `FormHandlerEntry` models. #37
+- Fixed element "name" field strips underscores issue. #33
+
 0.6.8
 -----
 2016-09-06